Backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
certifi==2025.7.14; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cython==0.29.21; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
dnspython==2.7.0; python_version >= '3.9'
docopt==0.6.2
email-validator==2.2.0; python_version >= '3.8'
fastapi==0.116.1; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
numpy==2.3.1; python_version >= '3.11'
pandas==2.3.1; python_version >= '3.9'
passlib==1.7.4
pipreqs==0.4.13; python_version >= '3.7'
pydantic[email]==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
pytz==2025.2
requests==2.32.4; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.47.1; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
uvicorn==0.35.0; python_version >= '3.9'
yarg==0.1.10
